refactor: 리뷰 컨트롤러 분리 및 불필요한 음식점 아이디 파라미터 삭제 (#316)

+++ backend/src/main/java/com/celuveat/restaurant/application/RestaurantReviewService.java
@@ -31,12 +31,12 @@ public class RestaurantReviewService {
 
     public void update(UpdateReviewRequestCommand command) {
         RestaurantReview review = restaurantReviewRepository.getById(command.reviewId());
-        review.updateContent(command.content(), command.memberId(), command.restaurantId());
+        review.updateContent(command.content(), command.memberId());
     }
 
     public void delete(DeleteReviewCommand command) {
         RestaurantReview review = restaurantReviewRepository.getById(command.reviewId());
-        review.checkMemberAndRestaurantMatched(command.memberId(), command.restaurantId());
+        review.checkOwner(command.memberId());
         restaurantReviewRepository.delete(review);
     }
 }
